tg_bot/requirements.txt
python-telegram-bot==12.6.1
mysql-connector-python==8.0.19
